+++ src/js/layout/displayMediaQueries/Slogan .jsx
