--- src/js/layout/main layout/routes/Error404.jsx
@@ -1,14 +1,0 @@
-// OK
-
-import React from 'react';
-
-const Error404 = () => {
-
-    return (
-        <main className="outlet--container">
-            <h2>404, please turn back.</h2>
-        </main>
-    );
-};
-
-export default Error404;
