--- src/js/layout/main layout/routes/HomePage.jsx
@@ -1,11 +1,0 @@
-//ok
-
-import React from 'react';
-
-const HomePage = () => {
-    return (
-        <main>Home</main>
-    );
-};
-
-export default HomePage;
