+++ src/js/layout/routes/Fave.jsx
@@ -1,0 +1,16 @@
+//ok
+
+import React from 'react';
+import FaveSearch from "../../components/FaveSearchComponents/FaveSearch.jsx"
+
+
+const Fave = () => {
+    return (
+        <main className="outlet--container">
+            <section className="main__section">
+                <FaveSearch />
+            </section>
+        </main>
+)};
+
+export default Fave;
