+++ src/js/layout/routes/HomePage.jsx
@@ -1,0 +1,15 @@
+//ok
+
+import React from 'react';
+
+const HomePage = () => {
+    return (
+        <main className="outlet--container">
+            <section className="main__section">
+                <h2>Welcome!</h2>
+            </section>
+        </main>
+    );
+};
+
+export default HomePage;
